+++ test/Fixture/Files/DeclareStrictTypesRule/Failure/file-with-comment-and-declare-strict-types-off-and-invalid-casing.php
@@ -1,0 +1,7 @@
+<?php
+
+// Hmm
+
+declare(sTrIcT_tYpEs=0);
+
+echo 'Hello!';