+++ test/Fixture/Files/DeclareStrictTypesRule/Failure/file-with-declare-ticks.php
@@ -1,0 +1,5 @@
+<?php
+
+declare(ticks=5);
+
+echo 'Hello!';