+++ test/Fixture/Files/DeclareStrictTypesRule/Success/file-with-comment-and-declare-strict-types-on-and-invalid-casing.php
@@ -1,0 +1,7 @@
+<?php
+
+// Hmm
+
+declare(sTrIcT_tYpEs=1);
+
+echo 'Hello!';