+++ test/Fixture/Files/DeclareStrictTypesRule/Success/file-with-comment-and-declare-strict-types-on-and-multiple-declares.php
@@ -1,0 +1,7 @@
+<?php
+
+// Hmm
+
+declare(ticks=5, strict_types = 1);
+
+echo 'Hello!';