+++ test/Fixture/Files/DeclareStrictTypesRule/Success/file-with-declare-strict-types-on-and-multiple-declares.php
@@ -1,0 +1,5 @@
+<?php
+
+declare(ticks=5, strict_types = 1);
+
+echo 'Hello!';