+++ test/Fixture/Files/DeclareStrictTypesRule/Success/file-with-doc-block-and-declare-strict-types-on-and-namespace-declaration.php
@@ -1,0 +1,11 @@
+<?php
+
+/**
+ * Hmm.
+ */
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Classes\DeclareStrictTypesRule\Success;
+
+echo 'Hello!';